Reworked to use timer and pulse interrupt
The original did not work for me as I am using a TSOP 7000 which is
active low. I am not sure how the pulses aredetected in the original.
The new version uses a timer interrupt to measure the delay between
pulses, and an interrupt on the trailing edge of the pulse to process
the delays and figure out the data payload.